app/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
openai==1.61.0
python-dotenv==1.0.1
